+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `time_us`
 * updated `stringify` to correctly handle None-cases
+* added `satella.cassandra.parallel_for`